fix: Change server directory definition to look for serverpod in pubspec instead of generate.yaml file.
