+++ tools/serverpod_cli/lib/src/util/directory.dart
@@ -1,6 +1,7 @@
 import 'dart:io';
 
 import 'package:path/path.dart' as p;
+import 'package:serverpod_cli/src/util/pubspec_helpers.dart';
 
 bool isServerpodRootDirectory([Directory? directory]) {
   // Verify that we are in the serverpod directory
@@ -20,11 +21,25 @@ bool isServerpodRootDirectory([Directory? directory]) {
 
 bool isServerDirectory(Directory directory) {
   var pubspec = File(p.join(directory.path, 'pubspec.yaml'));
-  var generator = File(p.join(directory.path, 'config', 'generator.yaml'));
 
-  if (!pubspec.existsSync() || !generator.existsSync()) {
-    return false;
-  }
+  if (!pubspec.existsSync()) return false;
+
+  var content = parsePubspec(pubspec);
+  if (!content.dependencies.containsKey('serverpod')) return false;
 
   return true;
+}
+
+Directory? findServerDirectory(Directory root) {
+  if (isServerDirectory(root)) return root;
+
+  var childDirs = root.listSync().where(
+        (dir) => isServerDirectory(Directory.fromUri(dir.uri)),
+      );
+
+  if (childDirs.isNotEmpty) {
+    return Directory.fromUri(childDirs.first.uri);
+  }
+
+  return null;
 }
